Restructure of files for manual grading pages
