+++ pages/instructorAssessmentManualGrading/assessment/assessment.ejs
@@ -1,16 +1,16 @@
 <!DOCTYPE html>
 <html lang="en">
   <head>
-    <%- include('../partials/head'); %>
+    <%- include('../../partials/head'); %>
     <link href="<%= node_modules_asset_path('tablesorter/dist/css/theme.bootstrap.min.css') %>" rel="stylesheet" />
     <script src="<%= node_modules_asset_path('tablesorter/dist/js/jquery.tablesorter.min.js') %>"></script>
     <script src="<%= node_modules_asset_path('tablesorter/dist/js/jquery.tablesorter.widgets.min.js') %>"></script>
   </head>
   <body>
-    <%- include('../partials/navbar'); %>
+    <%- include('../../partials/navbar'); %>
     <div id="content" class="container-fluid">
-      <%- include('../partials/assessmentSyncErrorsAndWarnings'); %>
-      <%- include('../partials/assessmentOpenInstancesAlert') %>
+      <%- include('../../partials/assessmentSyncErrorsAndWarnings'); %>
+      <%- include('../../partials/assessmentOpenInstancesAlert') %>
       <div class="card mb-4">
         <div class="card-header bg-primary text-white">
           <%= assessment_set.name %> <%= assessment.number %>: Manual Grading Queue
