+++ pages/instructorAssessmentManualGrading/assessment/assessment.js
@@ -3,8 +3,8 @@ const router = express.Router();
 const asyncHandler = require('express-async-handler');
 const path = require('path');
 const debug = require('debug')('prairielearn:' + path.basename(__filename, '.js'));
-const sqlDb = require('../../prairielib/lib/sql-db');
-const sqlLoader = require('../../prairielib/lib/sql-loader');
+const sqlDb = require('../../../prairielib/lib/sql-db');
+const sqlLoader = require('../../../prairielib/lib/sql-loader');
 
 const sql = sqlLoader.loadSqlEquiv(__filename);
